pyproject.toml
[project]
name = "ai-npc-validator"
version = "0.1.0"
description = "验证小型 LLM 模型在 Windows CPU 上的端侧推理性能"
requires-python = ">=3.9"
dependencies = [
    "airtest>=1.4.3",
    "google-genai>=1.47.0",
    "llama-cpp-python>=0.2.0",
    "poco>=0.99.3",
    "pocoui>=1.0.94",
    "requests>=2.31.0",
]


